backend/requirements.txt
amqp==5.3.1
appdirs==1.4.4
APScheduler==3.11.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
bs4==0.0.2
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
cssselect==1.2.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fake-useragent==2.0.3
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
kombu==5.5.2
lxml==5.3.1
lxml_html_clean==0.4.1
msgpack==1.1.0
parse==1.20.2
playwright==1.50.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyee==12.1.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyppeteer==0.0.25
pyquery==2.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-html==0.10.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.3.8
urllib3==1.26.20
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.13
websockets==10.4
whitenoise==6.9.0
zipp==3.21.0
zope.interface==7.2
